+++ lib/WorseReflection/Core/Inference/Resolver/BinaryExpressionResolver.php
@@ -167,6 +167,11 @@ class BinaryExpressionResolver implements Resolver
         Node $rightOperand,
         int $operator
     ): NodeContext {
+        switch ($operator) {
+            case TokenKind::OrKeyword:
+            case TokenKind::BarBarToken:
+
+        }
         if (!NodeUtil::canAcceptTypeAssertion($leftOperand, $rightOperand)) {
             return $context;
         }
